accomodate the preferred standalone function invocation pattern
The goal here is to allow the functions to be invoked in two ways.
First with an object reciever that wraps the element to be manipulated
so long as it provides it as the .e property. Second with the additional
dom element as a second argument and the dom element to be manipulated as
the first. Time will tell if this pattern will work across the board.