simplify the argument handling in the methods themselves
The goal with this commit is to push the argument handling required
by invocation othe Wield.Dom object out of the methods themselves.
The methods should be the simplest possible implementation. This also
allows for composition of the methods as they can return the element
they are operating on where appropriate.